+++ packages/expo-router/src/layouts/experimental-stack/createExperimentalStackNavigator.tsx
@@ -17,6 +17,8 @@ import type {
 } from './types';
 
 export interface ExperimentalStackNavigatorCreateProps {
+  isPreloaded: (key: string) => boolean;
+  isRemovalPrevented: (key: string) => boolean;
   pop: (count: number, sourceRouteKey: string) => void;
   removeRoutes: (routeNames: string[]) => void;
   subscribePopToTopOnParentTabPress: () => (() => void) | undefined;
@@ -39,6 +41,8 @@ function ExperimentalStackNavigatorContent({
   state,
   descriptors,
   emitter,
+  isPreloaded,
+  isRemovalPrevented,
   pop,
   removeRoutes,
   subscribePopToTopOnParentTabPress,
@@ -58,6 +62,8 @@ function ExperimentalStackNavigatorContent({
       <ExperimentalStackView
         state={state}
         emit={emitter.emit}
+        isPreloaded={isPreloaded}
+        isRemovalPrevented={isRemovalPrevented}
         pop={pop}
         descriptors={mergedDescriptors}
       />